Preserve information

+++ src/ripple/overlay/impl/PeerReservationTable.cpp
@@ -20,13 +20,27 @@
 #include <ripple/overlay/PeerReservationTable.h>
 
 #include <ripple/core/DatabaseCon.h>
+#include <ripple/json/json_value.h>
 #include <ripple/protocol/PublicKey.h>
+#include <ripple/protocol/jss.h>
 
 #include <boost/optional.hpp>
 
 #include <string>
 
 namespace ripple {
+
+auto
+PeerReservation::toJson() const -> Json::Value
+{
+    Json::Value result{Json::objectValue};
+    result[jss::node] = toBase58(TokenType::PublicKey, nodeId_);
+    if (description_)
+    {
+        result[jss::description] = description_;
+    }
+    return result;
+}
 
 // See `ripple/app/main/DBInit.cpp` for the `CREATE TABLE` statement.
 // REVIEWER: It is unfortunate that we do not get to define a function for it.
@@ -52,12 +66,13 @@ PeerReservationTable::load(DatabaseCon& connection)
     st.execute();
     while (st.fetch())
     {
-        if (!valPubKey) {
+        if (!valPubKey)
+        {
             // REVIEWER: How to signal unreachable? This represents a `NULL`
             // in a `NOT NULL` column.
         }
-        auto const optNodeId = parseBase58<PublicKey>(
-            TokenType::NodePublic, *valPubKey);
+        auto const optNodeId =
+            parseBase58<PublicKey>(TokenType::NodePublic, *valPubKey);
         if (!optNodeId)
         {
             // REVIEWER: Does the call site filter the level?
@@ -75,38 +90,53 @@ PeerReservationTable::load(DatabaseCon& connection)
     return true;
 }
 
-bool PeerReservationTable::upsert(
-        PublicKey const& nodeId,
-        boost::optional<std::string> const& desc
-)
+auto
+PeerReservationTable::upsert(
+    PublicKey const& nodeId,
+    boost::optional<std::string> const& desc)
+    -> boost::optional<PeerReservation>
 {
-    auto emplaced = table_.emplace(nodeId, PeerReservation{nodeId, desc});
+    PeerReservation const rvn{nodeId, desc};
+    boost::optional<PeerReservation> previous;
+
+    auto emplaced = table_.emplace(nodeId, rvn);
     if (!emplaced.second)
+    {
         // The node already has a reservation.
         // I think most people just want to overwrite existing reservations.
         // TODO: Make a formal team decision that we do not want to raise an
         // error upon a collision.
-        emplaced.first->second.description_ = desc;
+        previous = emplaced.first->second;
+        emplaced.first->second = rvn;
+    }
 
     auto db = connection_->checkoutDb();
     *db << "INSERT INTO PeerReservations (PublicKey, Description) "
-        "VALUES (:nodeId, :desc) "
-        "ON CONFLICT (PublicKey) DO UPDATE SET Description=excluded.Description",
-        soci::use(toBase58(TokenType::NodePublic, nodeId)),
-        soci::use(desc);
+           "VALUES (:nodeId, :desc) "
+           "ON CONFLICT (PublicKey) DO UPDATE SET "
+           "Description=excluded.Description",
+        soci::use(toBase58(TokenType::NodePublic, nodeId)), soci::use(desc);
 
-    return emplaced.second;
+    return previous;
 }
 
-bool PeerReservationTable::erase(PublicKey const& nodeId)
+auto
+PeerReservationTable::erase(PublicKey const& nodeId)
+    -> boost::optional<PeerReservation>
 {
-    bool const removed = table_.erase(nodeId) > 0;
+    boost::optional<PeerReservation> previous;
 
-    auto db = connection_->checkoutDb();
-    *db << "DELETE FROM PeerReservations WHERE PublicKey = :nodeId",
-        soci::use(toBase58(TokenType::NodePublic, nodeId));
+    auto it const = table_.find(nodeId);
+    if (it != table_.end())
+    {
+        previous = it->second;
+        table_.erase(it);
+        auto db = connection_->checkoutDb();
+        *db << "DELETE FROM PeerReservations WHERE PublicKey = :nodeId",
+            soci::use(toBase58(TokenType::NodePublic, nodeId));
+    }
 
-    return removed;
+    return previous;
 }
 
 }  // namespace ripple
